# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/video/formData":{"post":{"summary":"Upload Video Formdata","operationId":"upload_video_formData_video_formData_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_video_formData_video_formData_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/video/url":{"post":{"summary":"Submit Video Url","operationId":"submit_video_url_video_url_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoURL"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/video/{process_id}":{"get":{"summary":"Get Video Status","operationId":"get_video_status_video__process_id__get","parameters":[{"name":"process_id","in":"path","required":true,"schema":{"type":"string","title":"Process Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_video_formData_video_formData_post":{"properties":{"video":{"type":"string","format":"binary","title":"Video"}},"type":"object","required":["video"],"title":"Body_upload_video_formData_video_formData_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"StatusResponse":{"properties":{"status":{"type":"string","title":"Status"},"tags":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Tags"}},"type":"object","required":["status"],"title":"StatusResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VideoResponse":{"properties":{"id":{"type":"string","title":"Id"}},"type":"object","required":["id"],"title":"VideoResponse"},"VideoURL":{"properties":{"url":{"type":"string","maxLength":2083,"minLength":1,"format":"uri","title":"Url"}},"type":"object","required":["url"],"title":"VideoURL"}}}}